requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyTelegramBotAPI==4.24.0
python-telegram-bot==21.8
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
